--- src/types/EnrollmentForm/AdditionalInfo.ts
@@ -1,92 +1,0 @@
-import { z } from "zod";
-
-export const additionalInfoValidator = z.object({
-  rigsWithoutCigs: z.object({
-    tobaccoForm: z.object({
-      doesUseCigarettes: z.boolean(),
-      doesUseSmokelessTobacco: z.boolean(),
-    }),
-    tobaccoUsageLength: z.string(),
-    isFirstTimeTryingToQuit: z.boolean(),
-    methodsUsedToQuit: z.object({
-      hasNotTriedToQuit: z.boolean(),
-      hasTriedColdTurkey: z.boolean(),
-      hasUsedAudiobook: z.boolean(),
-      hasUsedChantix: z.boolean(),
-      hasUsedColdTurkey: z.boolean(),
-      hasUsedECigarettes: z.boolean(),
-      hasUsedGrindsCoffeePouches: z.boolean(),
-      hasUsedGum: z.boolean(),
-      hasUsedHypnosis: z.boolean(),
-      hasUsedLozenges: z.boolean(),
-      hasUsedMedication: z.boolean(),
-      hasUsedMobileApp: z.boolean(),
-      hasUsedNicotinePatch: z.boolean(),
-      hasUsedOther: z.boolean(),
-      hasUsedTaperMethod: z.boolean(),
-      hasUsedVarenicline: z.boolean(),
-      hasUsedWellbutrin: z.boolean(),
-    }),
-    tobaccoFagerstromScore: z.number(),
-    firstTobaccoTime: z.enum([
-      "Within 5 minutes",
-      "6-30 minutes",
-      "31-60 minutes",
-      "After 60 minutes",
-    ]),
-    swallowTobaccoJuice: z.enum(["Always", "Sometimes", "Never"]),
-    tobaccoHateToGiveUp: z.enum(["The first one in the morning", "All others"]),
-    tobaccoCansPerWeek: z.enum(["More than 3", "2-3", "1"]),
-    tobaccoChewMoreAfterAwakening: z.boolean(),
-    tobaccoChewWhenIll: z.boolean(),
-    cigaretteFagerstromScore: z.number(),
-    firstSmokeTime: z.enum([
-      "Within 5 minutes",
-      "6-30 minutes",
-      "31-60 minutes",
-      "After 60 minutes",
-    ]),
-    isDifficultToNotSmokeInForbiddenAreas: z.boolean(),
-    cigaretteHateToGiveUp: z.enum([
-      "The first one in the morning",
-      "All others",
-    ]),
-    cigarettesPerDay: z.enum(["31 or more", "21-30", "11-20", "10 or less"]),
-    smokeMoreInMorning: z.boolean(),
-    smokeWhenIll: z.boolean(),
-    plansToJoinFacebookGroup: z.boolean(),
-    whyDoYouWantToQuitSmoking: z.string(),
-    howCanWeHelpYou: z.string(),
-    referralSource: z.enum([
-      "Website",
-      "Friend",
-      "Social Media",
-      "SCF",
-      "Radio-Dave Nemo",
-      "Radio-Tim Ridley Show",
-      "Radio-KC Phillips-Road Dog Live",
-      "Radio-Erice Harley-Red Eye Radio",
-      "Radio-Other",
-      "Prime Inc",
-      "Rolling Strong",
-      "AAOO",
-      "Enrollment Form prompt",
-      "Other",
-    ]),
-    accountabilityPerson: z.object({
-      firstName: z.string(),
-      lastName: z.string(),
-      phoneNumber: z.string(),
-      relationshipToAccountabilityPerson: z.enum([
-        "Friend",
-        "Coworker",
-        "Spouse",
-        "Significant other",
-        "Other",
-      ]),
-    }),
-    currentlyHasPrimaryCarePhysician: z.boolean(),
-  }),
-});
-
-export type AdditionalInfo = z.infer<typeof additionalInfoValidator>;
