+++ src/types/EnrollmentForm/index.ts
@@ -1,4 +1,3 @@
-export * from "./AdditionalInfo";
 export * from "./EnrollmentForm";
 export * from "./GeneralInformationSection";
 export * from "./ProgramSelectionSection";
